requirements.txt
google-generativeai
Pillow
numpy
streamlit>=1.28.0
python-dotenv